UPLATEAUInstancedCityModelImpを廃止しFPLATEAUAutomationTestBaseで内部関数化
