+++ Source/PLATEAUTests/Tests/PLATEAUAutomationTestBase.h
@@ -5,13 +5,59 @@
 #include "PLATEAUCityModelLoader.h"
 #include "PLATEAUImportModelBtn.h"
 #include "PLATEAUInstancedCityModel.h"
-#include "PLATEAUInstancedCityModelImp.h"
 #include "Kismet/GameplayStatics.h"
 #include "Misc/AutomationTest.h"
+#include "ExtentEditor/PLATEAUExtentGizmo.h"
+#include "PLATEAUEditor/Public/PLATEAUEditor.h"
+#include "PLATEAUEditor/Public/ExtentEditor/PLATEAUExtentEditor.h"
+#include <plateau/dataset/dataset_source.h>
 
 
-class FPLATEAUAutomationTestBase : public FAutomationTestBase
-{
+class FPLATEAUAutomationTestBase : public FAutomationTestBase {
+    struct FGizmoData {
+        FGizmoData(): MinX(0), MinY(0), MaxX(0), MaxY(0) {
+        }
+
+        FGizmoData(const double InMinX, const double InMinY, const double InMaxX, const double InMaxY): MinX(InMinX), MinY(InMinY), MaxX(InMaxX),
+            MaxY(InMaxY) {
+        }
+
+        double MinX;
+        double MinY;
+        double MaxX;
+        double MaxY;
+    };
+
+    APLATEAUCityModelLoader* GetLocalCityModelLoader(const int ZoneId, const FVector& ReferencePoint, const int64 PackageMask, const FString& SourcePath, const FGizmoData& GizmoData, const TMap<int64, FPackageInfoSettings>& PackageInfoSettingsData) const {
+        const auto& ExtentEditor = IPLATEAUEditorModule::Get().GetExtentEditor();
+        ExtentEditor->SetImportFromServer(false);
+        ExtentEditor->SetSourcePath(SourcePath);
+        ExtentEditor->SetLocalPackageMask(static_cast<plateau::dataset::PredefinedCityModelPackage>(PackageMask));
+    
+        const auto DatasetSource = plateau::dataset::DatasetSource::createLocal(TCHAR_TO_UTF8(*SourcePath));
+        const std::shared_ptr<plateau::dataset::IDatasetAccessor> DatasetAccessor = DatasetSource.getAccessor();
+        if (DatasetAccessor == nullptr || DatasetAccessor->getMeshCodes().size() == 0)
+            return nullptr;
+    
+        const plateau::geometry::GeoReference RawGeoReference(ZoneId, {}, 1, plateau::geometry::CoordinateSystem::ESU);
+        ExtentEditor->SetGeoReference(RawGeoReference);
+
+        auto GeoReference = ExtentEditor->GetGeoReference();
+        const auto RawCenterPoint = DatasetAccessor->calculateCenterPoint(GeoReference.GetData());
+        GeoReference.ReferencePoint.X = RawCenterPoint.x;
+        GeoReference.ReferencePoint.Y = RawCenterPoint.y;
+        GeoReference.ReferencePoint.Z = RawCenterPoint.z;
+        ExtentEditor->SetGeoReference(RawGeoReference);
+
+        const auto ExtentGizmo = MakeUnique<FPLATEAUExtentGizmo>();
+        ExtentGizmo->SetMaxX(GizmoData.MaxX);
+        ExtentGizmo->SetMaxY(GizmoData.MaxY);
+        ExtentGizmo->SetMinX(GizmoData.MinX);
+        ExtentGizmo->SetMinY(GizmoData.MinY);
+        ExtentEditor->SetExtent(ExtentGizmo->GetExtent(GeoReference));
+
+        return UPLATEAUImportModelBtn::GetCityModelLoader(ZoneId, ReferencePoint, PackageInfoSettingsData, false);
+    }
 public:
 	FPLATEAUAutomationTestBase(const FString& InName, const bool bInComplexTask)
 		: FAutomationTestBase(InName, bInComplexTask) {
@@ -46,7 +92,7 @@ protected:
             const FPackageInfoSettings PackageInfoSettings(true, true, 0, 4, 1);
             TMap<int64, FPackageInfoSettings> PackageInfoSettingsData;
             PackageInfoSettingsData.Add(static_cast<int64>(plateau::dataset::PredefinedCityModelPackage::Building), PackageInfoSettings);
-            const auto& Loader = UPLATEAUInstancedCityModelImp::GetLocalCityModelLoader(ZoneId, ReferencePoint, PackageMask, SourcePath, GizmoData, PackageInfoSettingsData);
+            const auto& Loader = GetLocalCityModelLoader(ZoneId, ReferencePoint, PackageMask, SourcePath, GizmoData, PackageInfoSettingsData);
             if (Loader) {
                 return Loader;
             }
